Refresh specs for Hash#replace
This includes a new spec to see it compare_by_identity is transfered,
this test is no longer needed in test/natalie/hash_test
